backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
openai
transformers
torch
